Integrità e validazione
Implementati controlli di integrità nelle classi service ed aggiunte le annotazioni necessarie alla validazione

+++ registro/src/main/java/com/psss/registro/backend/services/CrudService.java
@@ -4,6 +4,7 @@ import com.psss.registro.backend.models.AbstractEntity;
 import com.psss.registro.backend.models.Docente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import javax.persistence.EntityExistsException;
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
@@ -12,7 +13,15 @@ public interface CrudService <T extends AbstractEntity> {
 
     JpaRepository<T, Long> getRepository();
 
-    default T save(T entity){return getRepository().saveAndFlush(entity);}
+    default boolean save(T entity){
+        if(entity.getId() == null){
+            getRepository().saveAndFlush(entity);
+            return true;
+        }else{
+            return false;
+            //throw new EntityExistsException();
+        }
+    }
 
     default List<T> findAll() {
         return getRepository().findAll();
@@ -22,8 +31,14 @@ public interface CrudService <T extends AbstractEntity> {
         return getRepository().findById(id);
     }
 
-    default void deleteById(Long id){
+    default boolean deleteById(Long id){
+        T entity = getRepository().findById(id).orElse(null);
+        if(entity == null){
+            return false;
+            //throw EntityNotFoundException();
+        }
         getRepository().deleteById(id);
+        return true;
     }
 
 }
